deprecated/hdfs-kafka-consumer/requirements.txt
kafka-python
hdfs[avro]
pandas
pyarrow
